+++ tests/files/src/0435_false_positive_shape_offset.php
@@ -5,3 +5,5 @@ function exampleShape436($x) {
     echo intdiv($x['key'], 2);  // should warn
     return $x['other']; // Should not warn.
 }
+$global435['xyz'] = 'value';
+echo count($global435['differentKey']);  // should assume this was set somewhere else